requirements.txt
accelerate==0.31.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.11.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
Babel==2.15.0
beartype==0.18.5
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.1.0
cachetools==5.3.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.3
colorama==0.4.6
colorful==0.5.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
datasets==2.21.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
diffusers @ git+https://github.com/huggingface/diffusers.git@fbe29c62984c33c6cf9cf7ad120a992fe6d20854
dill==0.3.8
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
email_validator==2.1.1
-e git+ssh://git@github.com/codezakh/11-envgen-pp.git@46fdc5482fabe645bc21fc6026acd7405673490c#egg=envgenpp
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.112.2
fastapi-cli==0.0.4
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.15.4
fire==0.6.0
fonttools==4.52.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gguf==0.9.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.8.0
google-auth==2.29.0
googleapis-common-protos==1.56.1
gradio==4.31.5
gradio_client==0.16.4
grpcio==1.64.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
hydra-core @ git+https://github.com/facebookresearch/hydra.git@f4fe48442992defef7c2ddd0a6d014e3c371a073
hydra_zen==0.13.0
icecream==2.1.3
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.0
iniconfig==2.0.0
instructor==1.4.0
interegular==0.3.3
ipdb==0.13.13
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-kernel-gateway==3.0.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lark==1.2.2
lazy_loader==0.4
librosa==0.10.2.post1
linkify-it-py==2.0.3
-e git+ssh://git@github.com/codezakh/11-envgen-pp.git@46fdc5482fabe645bc21fc6026acd7405673490c#egg=llamafactory&subdirectory=src/external/LLaMA-Factory
llvmlite==0.43.0
lm-format-enforcer==0.10.6
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
memray==1.12.0
mistral_common==1.3.4
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
mypy==1.10.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
notebook==7.2.0
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
omegaconf @ git+https://github.com/omry/omegaconf@6f9656bf091f6cb1ebfe22c926e60e1c9d25ca02
openai==1.43.0
opencensus==0.11.4
opencensus-context==0.1.3
orjson==3.10.3
outlines==0.0.46
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post4
pathspec==0.12.1
peft==0.12.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pooch==1.8.2
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==5.28.0
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==9.0.0
py-spy==0.3.14
pyairports==2.1.1
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
pytest==8.2.1
pytest-forked==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-ulid==2.7.0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.5.2
QtPy==2.4.1
ray==2.35.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
ruff==0.4.4
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.12.0
setproctitle==1.3.3
sh==2.0.6
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.5.0.post1
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.38.4
sympy==1.13.2
tenacity==8.5.0
termcolor==2.4.0
terminado==0.18.1
textual==0.63.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
triton==3.0.0
trl @ git+https://github.com/huggingface/trl.git@ac7c8b128494dc98db3c795c728da3bf7abed981
typer==0.12.3
types-python-dateutil==2.9.0.20240316
types-tqdm==4.66.0.20240417
typing_extensions==4.12.2
tyro==0.8.5
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.2
uv==0.1.45
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.26.2
vllm==0.6.0
vllm-flash-attn==2.6.1
vllm_nccl_cu12==2.18.1.0.4.0
wandb==0.17.5
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0.1
widgetsnbextension==4.0.10
wrapt==1.16.0
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.9.11
zipp==3.20.1
